+++ Realm/Tests/RLMTestCase.h
@@ -25,6 +25,7 @@ extern "C" {
 NSString *RLMTestRealmPath(void);
 NSString *RLMDefaultRealmPath(void);
 NSString *RLMRealmPathForFile(NSString *);
+NSData *RLMGenerateKey(void);
 #ifdef __cplusplus
 }
 #endif
@@ -34,6 +35,6 @@ NSString *RLMRealmPathForFile(NSString *);
 - (RLMRealm *)realmWithTestPath;
 - (RLMRealm *)realmWithTestPathAndSchema:(RLMSchema *)schema;
 
-+ (void)deleteFiles;
+- (void)deleteFiles;
 
 @end