+++ Realm/Tests/RLMTestCase.m
@@ -69,25 +69,46 @@ static void RLMDeleteRealmFilesAtPath(NSString *path) {
     }
 }
 
+NSData *RLMGenerateKey() {
+    uint8_t buffer[64];
+    SecRandomCopyBytes(kSecRandomDefault, 64, buffer);
+    return [[NSData alloc] initWithBytes:buffer length:sizeof(buffer)];
+}
+
+static BOOL encryptTests() {
+    static BOOL encryptAll = NO;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        if (getenv("REALM_ENCRYPT_ALL")) {
+            encryptAll = YES;
+        }
+    });
+    return encryptAll;
+}
 
 @implementation RLMTestCase
 
-+ (void)setUp
+- (void)setUp
 {
     [super setUp];
     
     // Delete Realm files
     RLMDeleteRealmFilesAtPath(RLMDefaultRealmPath());
     RLMDeleteRealmFilesAtPath(RLMTestRealmPath());
+
+    if (encryptTests()) {
+        [RLMRealm setEncryptionKey:RLMGenerateKey() forRealmsAtPath:RLMDefaultRealmPath()];
+        [RLMRealm setEncryptionKey:RLMGenerateKey() forRealmsAtPath:RLMTestRealmPath()];
+    }
 }
 
-+ (void)tearDown
+- (void)tearDown
 {
     [super tearDown];
     [self deleteFiles];
 }
 
-+ (void)deleteFiles {
+- (void)deleteFiles {
     // Clear cache
     [RLMRealm resetRealmState];
     
@@ -98,11 +119,9 @@ static void RLMDeleteRealmFilesAtPath(NSString *path) {
 
 - (void)invokeTest
 {
-    [RLMTestCase setUp];
-    @autoreleasepool {
-        [super invokeTest];
-    }
-    [RLMTestCase tearDown];
+    @autoreleasepool { [self setUp]; }
+    @autoreleasepool { [super invokeTest]; }
+    @autoreleasepool { [self tearDown]; }
 }
 
 - (RLMRealm *)realmWithTestPath